# lib/api-zod/src/generated/types/dashboardData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HealthSync AI Health Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { Medication } from "./medication";
import type { SymptomEntry } from "./symptomEntry";
import type { WaterIntake } from "./waterIntake";

export interface DashboardData {
  todaysMedications: Medication[];
  waterIntake: WaterIntake;
  latestSymptoms?: SymptomEntry;
  /** @nullable */
  riskLevel?: string | null;
  weeklyExerciseCount: number;
  streak: number;
}