# lib/api-zod/src/generated/types/nutritionEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HealthSync AI Health Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface NutritionEntry {
  id: number;
  userId: string;
  date: string;
  meal: string;
  foods: string[];
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  protein?: number | null;
  /** @nullable */
  carbs?: number | null;
  /** @nullable */
  fat?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}