# lib/api-zod/src/generated/types/sOSBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HealthSync AI Health Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface SOSBody {
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  message: string;
}