# lib/api-zod/src/generated/types/waterIntake.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HealthSync AI Health Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface WaterIntake {
  /** @nullable */
  id?: number | null;
  userId: string;
  date: string;
  glasses: number;
  goalGlasses: number;
}